Fix int8 packet fields being decoded as unsigned bytes (#405)
## Summary

- `DRSAllowed`, `VehicleFiaFlags`/`ZoneFlag`, `Gear`, `TrackTemperature`
and `AirTemperature` are signed int8 fields per the F1 telemetry spec,
but were read as a raw unsigned byte and assigned without a sign cast.
The documented negative values (`-1` for unknown/invalid/reverse gear,
sub-zero temperatures) therefore arrived as 128-255 instead.
- Reads for these fields in `PacketToCarStatus`, `PacketToCarTelemetry`
and `PacketToSessionData` now go through an `sbyte` cast before
assignment so the sign is preserved.
- `TrackTemperature`/`AirTemperature` are widened from `ushort` to
`short` on `ISessionDataBase`, the F1 2019-2023 `SessionData` classes
and `WeatherForecastSample` so the negative value can actually be
represented.
- `TrackId` is a separate, already-tracked finding and is out of scope
here.

## Test plan

- [x] `dotnet build F1Server.Core/F1Server.Core.csproj` /
`F1Server.Tests/F1Server.Tests.csproj` - 0 warnings
- [x] `reihitsu-format --check` on all changed files - no formatting
needed
- [x] `dotnet test F1Server.Tests` - all 619 tests pass, including 5 new
regression tests covering `DRSAllowed`, `FiaFlags`, `Gear`,
`TrackTemperature`/`AirTemperature` and `ZoneFlag` with a synthetic byte
value of 255

Closes #78

+++ F1Server.Core/Packets/PacketToObject/PacketToCarStatus.cs
@@ -229,7 +229,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2019.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -330,7 +330,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2020.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -375,7 +375,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2021.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -424,7 +424,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2022.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -473,7 +473,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2023.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -522,7 +522,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2024.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -571,7 +571,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2025.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -620,7 +620,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
                     actOffset += ConstData.TypeUInt8;
 
-                    var flagsByte = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                    var flagsByte = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                     status2026.FiaFlags = (VehicleFiaFlagColor)Enum.ToObject(typeof(VehicleFiaFlagColor), flagsByte);
 
@@ -697,7 +697,7 @@ internal class PacketToCarStatus(PacketHeader packetHeader) : PacketToXBase(pack
 
             actOffset += ConstData.TypeUInt8;
 
-            carStatusData.DRSAllowed = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+            carStatusData.DRSAllowed = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             actOffset += ConstData.TypeUInt8;
         }
